Change the returned dict of function "calculate_model_scores()"
The function returns a dictionary with anomalies as keys; for each anomaly, the associated value
is the average number of the anomaly detections in the dataset (n_detection/dataset_lenght).
Before this commit, it worked the same for false positives.
Now the value associated with the key "false_positives" is the total number of false positives in the dataset.